+++ hdspin/utils/general_utils.h
@@ -3,7 +3,7 @@
 #ifndef GENERAL_UTILS_H
 #define GENERAL_UTILS_H
 
-int binary_vector_to_int(const int *config, const int N);
+long long binary_vector_to_int(const int *config, const int N);
 
 void flip_spin_(int *config, const int idx);
